console: add `retain-for` cmd line arg
