+++ Cargo.toml
@@ -2,9 +2,11 @@
 members = [
     "console",
     "console-subscriber",
-    "console-api"
+    "console-api",
+    "console-util"
 ]
 resolver = "2"
 
 [patch.crates-io]
 tokio = { git = 'https://github.com/zaharidichev/tokio', branch = 'zd/instrument-sleep' }
+
