+++ console-subscriber/Cargo.toml
@@ -15,6 +15,7 @@ tokio = { version = "^1.10", features = ["sync", "time", "macros", "tracing"]}
 tokio-stream = "0.1"
 thread_local = "1.1.3"
 console-api = { path = "../console-api", features = ["transport"]}
+console-util = { path = "../console-util"}
 tonic = { version = "0.5", features = ["transport"] }
 tracing-core = "0.1.18"
 tracing = "0.1.26"
