+++ console-util/Cargo.toml
@@ -1,0 +1,5 @@
+[package]
+name = "console-util"
+version = "0.1.0"
+authors = ["Eliza Weisman <eliza@buoyant.io>"]
+edition = "2018"
