+++ console-util/src/lib.rs
@@ -1,0 +1,3 @@
+mod duration;
+
+pub use duration::parse_duration;
