+++ console/Cargo.toml
@@ -10,6 +10,7 @@ repository = "https://github.com/tokio-rs/console"
 [dependencies]
 atty = "0.2"
 console-api = { path = "../console-api", features = ["transport"]}
+console-util = { path = "../console-util"}
 clap = "3.0.0-beta.2"
 tokio = { version = "1", features = ["full", "rt-multi-thread"]}
 tonic = { version = "0.5", features = ["transport"] }
