+++ console/src/tasks.rs
@@ -24,6 +24,7 @@ pub(crate) struct State {
     new_tasks: Vec<TaskRef>,
     current_task_details: DetailsRef,
     temporality: Temporality,
+    retain_for: Duration,
 }
 
 #[derive(Debug)]
@@ -60,7 +61,6 @@ pub(crate) struct Task {
     fields: Vec<Field>,
     formatted_fields: Vec<Vec<Span<'static>>>,
     stats: Stats,
-    completed_for: usize,
     target: Arc<str>,
     name: Option<Arc<str>>,
 }
@@ -84,6 +84,7 @@ pub(crate) struct Metadata {
 struct Stats {
     polls: u64,
     created_at: SystemTime,
+    dropped_at: Option<SystemTime>,
     busy: Duration,
     last_poll_started: Option<SystemTime>,
     last_poll_ended: Option<SystemTime>,
@@ -121,7 +122,12 @@ pub(crate) enum FieldValue {
 }
 
 impl State {
-    const RETAIN_COMPLETED_FOR: usize = 6;
+    pub(crate) fn new(retain_for: Duration) -> Self {
+        Self {
+            retain_for,
+            ..Default::default()
+        }
+    }
 
     pub(crate) fn last_updated_at(&self) -> Option<SystemTime> {
         self.last_updated_at
@@ -194,16 +200,14 @@ impl State {
             let formatted_fields = Field::make_formatted(styles, &mut fields);
             let id = task.id?.id;
             let stats = stats_update.remove(&id)?.into();
-            let mut task = Task {
+            let task = Task {
                 name,
                 id,
                 fields,
                 formatted_fields,
                 stats,
-                completed_for: 0,
                 target: meta.target.clone(),
             };
-            task.update();
             let task = Rc::new(RefCell::new(task));
             new_list.push(Rc::downgrade(&task));
             Some((id, task))
@@ -214,7 +218,6 @@ impl State {
             if let Some(task) = self.tasks.get_mut(&id) {
                 let mut t = task.borrow_mut();
                 t.stats = stats.into();
-                t.update();
             }
         }
     }
@@ -249,14 +252,20 @@ impl State {
             return;
         }
 
-        self.tasks.retain(|_, task| {
-            let mut task = task.borrow_mut();
-            if task.completed_for == 0 {
-                return true;
-            }
-            task.completed_for += 1;
-            task.completed_for <= Self::RETAIN_COMPLETED_FOR
-        })
+        if let Some(now) = self.last_updated_at {
+            let retain_for = self.retain_for;
+            self.tasks.retain(|_, task| {
+                let task = task.borrow();
+
+                task.stats
+                    .dropped_at
+                    .map(|d| {
+                        let dropped_for = now.duration_since(d).unwrap();
+                        retain_for > dropped_for
+                    })
+                    .unwrap_or(true)
+            })
+        }
     }
 
     // temporality methods
@@ -378,13 +387,6 @@ impl Task {
     pub(crate) fn self_wakes(&self) -> u64 {
         self.stats.self_wakes
     }
-
-    fn update(&mut self) {
-        let completed = self.stats.total.is_some() && self.completed_for == 0;
-        if completed {
-            self.completed_for = 1;
-        }
-    }
 }
 
 impl Details {
@@ -407,7 +409,15 @@ impl From<proto::tasks::Stats> for Stats {
             Duration::from_secs(secs) + Duration::from_nanos(nanos)
         }
 
-        let total = pb.total_time.map(pb_duration);
+        let created_at = pb
+            .created_at
+            .expect("task span was never created")
+            .try_into()
+            .unwrap();
+
+        let dropped_at: Option<SystemTime> = pb.dropped_at.map(|v| v.try_into().unwrap());
+        let total = dropped_at.map(|d| d.duration_since(created_at).unwrap());
+
         let poll_stats = pb.poll_stats.expect("task should have poll stats");
         let busy = poll_stats.busy_time.map(pb_duration).unwrap_or_default();
         let idle = total.map(|total| total - busy);
@@ -418,11 +428,8 @@ impl From<proto::tasks::Stats> for Stats {
             last_poll_started: poll_stats.last_poll_started.map(|v| v.try_into().unwrap()),
             last_poll_ended: poll_stats.last_poll_ended.map(|v| v.try_into().unwrap()),
             polls: poll_stats.polls,
-            created_at: pb
-                .created_at
-                .expect("task span was never created")
-                .try_into()
-                .unwrap(),
+            created_at,
+            dropped_at,
             wakes: pb.wakes,
             waker_clones: pb.waker_clones,
             waker_drops: pb.waker_drops,
